Improve tests, add a few new cases.
Makes it so the `to` field only points to a typeName, not a full type.

Also add some cases like for `as` being in the typename (used to fail
before). Note that the current fix in the grammar.js is to use ' as '
instead of 'as'. Not sure if this the ideal fix. Maybe make use of
grammar.word or maybe some precedence stuff?